insider-threat-detection-uba/requirements.txt
# Core ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0

# Anomaly Detection
pyod>=1.0.0

# Real-time Processing
kafka-python>=2.0.0
redis>=4.0.0
celery>=5.2.0

# Web Framework and Dashboard
flask>=2.1.0
flask-socketio>=5.1.0
plotly>=5.0.0
dash>=2.5.0
gunicorn>=20.1.0

# Database
sqlalchemy>=1.4.0

# Data Collection and Monitoring
psutil>=5.8.0
watchdog>=2.1.0
wmi>=1.5.1;platform_system=="Windows"
pywin32>=227;platform_system=="Windows"

# Configuration
pyyaml>=6.0
configparser>=5.2.0

# Logging and Monitoring
structlog>=22.1.0
prometheus-client>=0.14.0

# Networking and Security
requests>=2.27.0
cryptography>=37.0.0
paramiko>=2.11.0

# Utilities
python-dateutil>=2.8.0
click>=8.0.0
schedule>=1.1.0
joblib>=1.1.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
mock>=4.0.0

# Development
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Real-time Communication
websockets>=10.0
eventlet>=0.33.0

# File Processing
openpyxl>=3.0.0
xlsxwriter>=3.0.0